Add: <upload>, <import>, <export> directive
(php)
ConnectionManager.php
add: ImportData action request handling
DatabaseManager.php
add: process_result property in the responseArray, for data importing
process message
update: update function parameter allowUpdateToEmptyValue, this is for
import function, e.g a table pk is a number, a zero value is allow
ExcelManager.php
update: if empty() is true of a cell value, do not ignore to set it in
the excel engine object, reason same as above.
--------------------------

the upload/import function are tested and tried.